chat::server
Module replication

Definition of replication logic.

Structs

ReceiveEventsRequest

Represents request from the node to receive events with total sequence number from the specified range [from, to].

ReplicateRequest

Represents request to replicate event.

TotalSeqNumMsg

Represents message with the total sequence number on the sender node.

TotalSeqNumRequest

Represents request to get receiver total seq number.

Functions

get_replica_total_seq_num

Get total seq number on the replica with specified address.

replicate_client_request

request_replica_events_from_range

Request events with global sequence number in the range [from, to].